ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
pydantic==2.5.0
scikit-learn==1.3.2
lxml==4.9.3
xmltodict==0.13.0
aiofiles==23.2.1
python-multipart==0.0.6
python-magic==0.4.27
openai==1.30.1
python-dotenv==1.0.0